+++ apps/web/src/components/exercise-generation/exercise-preview-page.tsx
@@ -16,7 +16,7 @@ import {
   ExecutePromptStatus,
   useExecuteAIPrompt,
 } from './exercise-generation-wizard/execute-ai-prompt'
-import ExerciseGenerationLoadingSparkles from '@/assets-webkit/img/sparkles.svg'
+import {Skeleton} from '../skeleton'
 import { FaIcon } from '@/components/fa-icon'
 import { ModalWithCloseButton } from '@/components/modal-with-close-button'
 import { useInstanceData } from '@/contexts/instance-context'
@@ -280,26 +280,5 @@ export function ExercisePreviewPage({
         </div>
       </div>
     </ModalWithCloseButton>
-  )
-}
-
-export function Skeleton() {
-  return (
-    <div className="relative h-full w-full">
-      <div className="flex animate-pulse flex-col space-y-4">
-        <div className="h-4 w-2/4 rounded bg-gray-300"></div>
-        <div className="mt-6 h-4 w-3/4 rounded bg-gray-300"></div>
-        <div className="h-4 w-3/4 rounded bg-gray-300"></div>
-        <div className="mt-6 h-4 w-1/2 rounded bg-gray-300"></div>
-        <div className="h-4 w-1/2 rounded bg-gray-300"></div>
-        <div className="h-4 w-1/2 rounded bg-gray-300"></div>
-        <div className="h-4 w-1/2 rounded bg-gray-300"></div>
-        <div className="mt-6 h-4 w-1/4 rounded bg-gray-300"></div>
-        <div className="h-4 w-1/4 rounded bg-gray-300"></div>
-      </div>
-      <div className="absolute left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2 transform">
-        <ExerciseGenerationLoadingSparkles className="h-24 w-24 animate-pulse" />
-      </div>
-    </div>
   )
 }
