+++ apps/web/src/components/skeleton.ts
@@ -1,0 +1,22 @@
+import ExerciseGenerationLoadingSparkles from '@/assets-webkit/img/sparkles.svg'
+
+export function Skeleton() {
+  return (
+    <div className="relative h-full w-full">
+      <div className="flex animate-pulse flex-col space-y-4">
+        <div className="h-4 w-2/4 rounded bg-gray-300"></div>
+        <div className="mt-6 h-4 w-3/4 rounded bg-gray-300"></div>
+        <div className="h-4 w-3/4 rounded bg-gray-300"></div>
+        <div className="mt-6 h-4 w-1/2 rounded bg-gray-300"></div>
+        <div className="h-4 w-1/2 rounded bg-gray-300"></div>
+        <div className="h-4 w-1/2 rounded bg-gray-300"></div>
+        <div className="h-4 w-1/2 rounded bg-gray-300"></div>
+        <div className="mt-6 h-4 w-1/4 rounded bg-gray-300"></div>
+        <div className="h-4 w-1/4 rounded bg-gray-300"></div>
+      </div>
+      <div className="absolute left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2 transform">
+        <ExerciseGenerationLoadingSparkles className="h-24 w-24 animate-pulse" />
+      </div>
+    </div>
+  )
+}
